Make it possible to set random pdb name (Kinda hardcoded).
Add option to keep '[' and ']' characters on paths.
This is still not the best because 'Util.ResolvePath' throws an
exception when passing a path containing invalid characters for a path.
But is still good enough for our purposes of generating a unique pdb name
to be able to hot reload dll libraries.